ClaimRewardsTxDetails component testing
